Fix printing bug, add field strength parameter
Before this commit, with the default `nimg=1`, the initial state would
be the only one printed at the end. This was an indexing error, and has
been fixed, so that the proper number of images/image times are now
reported.

In addition, the external field strength parameter is now properly
implemented (in 2D, at least).