Touch up ising class, add event data handling
We just add some const constraints to the ising class to prevent memory
access bugs, and added event data handling. We're currently in the
beginning of a rewrite to default to performing the full suite of Ising
simulation calculations over a range of temperatures, so the API will
greatly change in the next day or two.