Add correlation fn, optimize E/M calcs
Previously we tabulated the total energy at every state as a sort of
error-check for our dU method --- if one method had a flaw, the other
should've as well (and indeed it did). This worked out, to some degree,
but we forgot to remove it (and thus our production code was slow). Now
we only calculate the total energy once, and utilize our (assumingly
correct) dU method to update the energy at each step. This can cause
errors to accumulate over sweeps, though!

We also add in a correlation function calculator, though the specific
means of calculating the correlation length is still to be formalized.