Add grid sanitization option, rewrite BC code
We shifted the boundary condition enforcement to the `get_spin`
function, which now uses a fast postive modulus function to handle
everything behind-the-scenes.

We also added a command line option to ask whether, for each temp, the
grid should be rewritten for a different equilibrium approach. This
option will increase the number of sweeps each temperature on average
will need.