Add image snapshotting, remove private sweeps var
This commit adds experimental support for image snapshotting using
libfmt. By using libfmt instead of <format> in C++20 (since compiler
support for the standard is not yet mature it seems), hopefully people
with other kinds of compilers can still build this project.

Additionally, the interface for the wolff class has been changed to
allow user specification of the number of sweeps (Monte Carlo steps) to
perform for each individual wolff::run() call. This way, we don't have
to keep messing with the class's state to adjust how many runs to
perform.