Add multithreaded U, M data recording, comments
This adds std::thread-based measurement of energy and magnetization (with
fluctuations). The current implementation is not entirely bug-free (there
have been some <rare> observed discrepancies in energy calculations).

Additionally, the class has been more thoroughly documented using
doxygen-style comments. The design of the wolff class is better of an
example of what our overall ising class should look like: setting a
backend (single-spin or wolff, or even SW and others), along with
different built-in "tests" that iterate with a Monte Carlo step
granularity. Hopefully the code will be in a suitable state for this
integration soon.