Update pthread indexing, cf calculations, filename input
This update fixed a prior bug in the pthread function payload
indexing, which previously incremented the temperature from tmin
to tmax using the given dt values provided on the command line.
This led to some garbage output, probably due to float -> int
rounding being imprecise, so instead we index using the integer
index, and recalculate the temperature when needed.

We also added correlation function calculations for any given state of
the grid to the wolff class, which is an improvement over the
implementation from the previous ising class, and will be the new
standard for the original ising class due to its efficiency and better
interface.

Last but not least, we added runtime output file name parsing, though
this will be retooled later to use boost::program_options, or some other
runtime configuration parsing system (xml, yaml, whatever).