Refactor experimental features into main class
This update migrates the implementation of the Wolff class from the
experimental subdirectory into the main program, allowing both
single-spin flips and cluster flips to be chosen at runtime (once per
object initialization!). Many improvements to the structure of the test
program (such as the sweeping algorithm now relying on checking the
number of spin flips per run) have been backported to the original
class.

Additionally, everything for the project has been moved into the ising
namespace, and the previous ising and wolff classes have been
implemented as the "simulation" class. This should hopefully make the
overall program easier to organize and maintain.