Wrap grid pointer in std::unique_ptr
This removes the need for an explicit dtor to free up our allocated grid
member variable, as we can just rely on object destruction to
automatically delete the array. This adds some more considerations we
should address in terms of copy and move semantics, but we can address
those later.