Add Julia rewrite
This adds a new set of source files to the fray, this time written in
Julia! This has several motivations: for one, it lets me get more
comfortable with the language; but it also allows me to help out some
former mentors of mine in terms of teaching aids / examples.

The source files have been split off into the cpp/ and jl/ directories,
but within each directory, things should be as usual.